[firrtl] Define comparison for LayerOp
Add a `std::optional<int> LayerOp::compare(const LayerOp &other)` member
function.  Layers are partially ordered, hence the use of `std::optional`
to represent the case of two layers not having an order.

This uses a convention (documented in the comments) that child layers are
greater than lower layers.  This is based on the rough idea that child
layers are "higher" than lower layers because they "stack on top of" their
parents.

h/t @darthscsi for instigating this line of thinking.

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>